CR * rename package to updaterini, new utils can be added with new commands to cli * remove os.Getwd works fine if use default descr name as default value for output dir
