+++ .gitignore
@@ -18,4 +18,4 @@
 # vendor/
 
 # Debug default files
-/cmd/serv_upd_gen/serv_update.json
+/cmd/updaterini/serv_update.json
